projects: Add example project for ADF4382
Add example projects for the adf4382 using stm32, mbed and xilinx
platforms. This also includes the example project for IIO as well.
The projects have been tested using the EV-ADF4382SD1Z and SDP-K1 for
the stm32 and mbed platforms. The xilinx platform has been tested
using a custom design on a Zed Board.

Signed-off-by: chegbeli <ciprian.hegbeli@analog.com>